Update style2.css
Changed background to look very modern, and almost like a newspaper. I made a grid to make it look more like a newspaper, but still need to move the footer to the bottom of the page, and links on the left side of the page underneath the preamble. Font is still great!