Update style2.css
Finally got the sidebar to go underneath the intro portion to give it a nice modern look. I changed the margins for the top of the preamble and the summary in order to give nice spacing between the sections. I also got the foot to be underneath the main container, as I wanted to give it a magazine/menu/newspaper look.