Update style.css
Made gradient backgrounds for each of the sections, aside from the aside portion at the bottom. I took the background off of the intro sections, and combined the intro and zen supporting sections together to make the backgrounds the same for the main sections of the page. I added extra padding for the summary section, as I wanted it lower and more in the middle, I made the footer centered with its own box within the zen supporting section. I made the box model for the footer a different gradient background as well to make it look better. Finally I got the aside portion to display a 3 grid system.